Rename the component directory to be `Figure` as well
