+++ components/Figure/isLimited.test.ts
