+++ components/Figure/isLimited.ts
