+++ components/Figure/types.ts
