+++ components/Figure/withWebp.test.ts
