+++ components/Figure/withWebp.ts
