--- components/Picture/index.ts
@@ -1,1 +1,0 @@
-export * from "./Picture";
