+++ pages/blog/architecture-which-sucks.mdx
@@ -1,4 +1,4 @@
-import { Figure } from "@components/Picture";
+import { Figure } from "@components/Figure";
 
 export const metadata = {
   title: "Грязная архитектура",
