+++ pages/blog/dont-push-2.mdx
@@ -1,4 +1,4 @@
-import { Figure } from "@components/Picture";
+import { Figure } from "@components/Figure";
 
 export const metadata = {
   title: "Не надо заставлять, надо автоматизировать",
