+++ pages/blog/make-error-messages-great-again.mdx
@@ -1,4 +1,4 @@
-import { Figure } from "@components/Picture";
+import { Figure } from "@components/Figure";
 
 export const metadata = {
   title: "Сообщение об ошибке, от которого не горит",
