+++ pages/blog/you-really-dont-need-redux-now.mdx
@@ -1,4 +1,4 @@
-import { Picture } from "@components/Picture";
+import { Figure } from "@components/Figure";
 
 export const metadata = {
   title: "От Редакса к хукам",
